requirements.txt
gunicorn
flask
gitpython
ipython
